tests/requirements.txt
Django
django-model-utils
factory-boy
sendgrid
tablets
